add: RequireGuildAttribute fix: set-space-notice-message 無法取消通知設定的問題 fix: 發送Youtube通知時，Discord遇到未知錯誤會跳無法取得DiscordCode導致跳出的問題 misc: s!nrc 現在會過濾已關台的直播 misc: 指令註冊方式

+++ Discord Stream Notify Bot/SharedService/Youtube/ReminderAction.cs
@@ -395,7 +395,7 @@ namespace Discord_Stream_Notify_Bot.SharedService.Youtube
                     }
                     catch (Discord.Net.HttpException httpEx)
                     {
-                        if (httpEx.DiscordCode.Value == DiscordErrorCode.InsufficientPermissions || httpEx.DiscordCode.Value == DiscordErrorCode.MissingPermissions)
+                        if (httpEx.DiscordCode.HasValue && (httpEx.DiscordCode.Value == DiscordErrorCode.InsufficientPermissions || httpEx.DiscordCode.Value == DiscordErrorCode.MissingPermissions))
                         {
                             Log.Warn($"Youtube 通知 - 遺失權限 {item.GuildId} / {item.DiscordChannelId}");
                             db.NoticeYoutubeStreamChannel.RemoveRange(db.NoticeYoutubeStreamChannel.Where((x) => x.DiscordChannelId == item.DiscordChannelId));
